make the LED blink when connecting to wifi
Former-commit-id: f51b2cef0cdaf90634ab3dca934adae3e1738c08
